backend/app/requirements.txt
fastapi==0.115.0
fastapi-cli==0.0.5
libvirt-python==10.7.0
